fix: converting text alignment caused columns to disappear
